assignments/assignment1/requirements.txt
matplotlib==3.6.2
numpy==1.26.4
pandas==1.5.0
scikit_learn==1.1.3
opencv-python-headless==4.9.0.80
tensorflow-hub
tensorflow-cpu
tqdm==4.66.2
